LC-distinct-subsequences [Runtime: N/A; Memory: N/A]

+++ dynamic-programming/distinct-subsequences.cpp
@@ -5,7 +5,7 @@ public:
         int n = s.size();
         int m = t.size();
 
-        vector<vector<int>> dp(n+1, vector<int>(m+1, 0));
+        vector<vector<long long>> dp(n+1, vector<long long>(m+1, 0));
         for(int i=0; i<=n; i++){
             dp[i][0] = 1;
         }
